+++ examples/singularity/singularity-exp.py
@@ -3,19 +3,25 @@
 """
 Example experiment for running Singularity planner images.
 
-Note that Downward Lab assumes that the evaluated algorithms are written
-in good faith. It is not equipped to handle malicious code. For example,
-it would be easy to write planner code that bypasses the time and memory
-limits set within Downward Lab. If you're running untrusted code, we
-recommend using cgroups to enforce resource limits.
+The time and memory limits set with Lab can be circumvented by solvers
+that fork child processes. Their resource usage is not checked. If you're
+running solvers that don't check their resource usage like Fast Downward,
+we recommend using cgroups or the "runsolver" tool to enforce resource
+limits. Since setting time limits for solvers with cgroups is difficult,
+the experiment below uses the "runsolver" tool, which has been used in
+multiple SAT competitions to enforce resource limits. For the experiment
+to run, the runsolver binary needs to be on the PATH. You can obtain a
+runsolver copy from https://github.com/jendrikseipp/runsolver.
 
-A note on running Singularity on clusters: reading large Singularity
-files over the network is not optimal, so we recommend copying the
-images to a local filesystem (e.g., /tmp/) before running experiments.
+A note on running Singularity on clusters: reading large Singularity files
+over the network is not optimal, so we recommend copying the images to a
+local filesystem (e.g., /tmp/) before running experiments.
 """
 
 import os
+from pathlib import Path
 import platform
+import sys
 
 from downward import suites
 from downward.reports.absolute import AbsoluteReport
@@ -38,11 +44,12 @@ class BaseReport(AbsoluteReport):
 
 NODE = platform.node()
 RUNNING_ON_CLUSTER = NODE.endswith((".scicore.unibas.ch", ".cluster.bc2.ch"))
-DIR = os.path.abspath(os.path.dirname(__file__))
-REPO = os.path.dirname(DIR)
-IMAGES_DIR = os.environ["SINGULARITY_IMAGES"]
-assert os.path.isdir(IMAGES_DIR), IMAGES_DIR
+DIR = Path(__file__).resolve().parent
+REPO = DIR.parent
+IMAGES_DIR = Path(os.environ["SINGULARITY_IMAGES"])
+assert IMAGES_DIR.is_dir(), IMAGES_DIR
 BENCHMARKS_DIR = os.environ["DOWNWARD_BENCHMARKS"]
+MEMORY_LIMIT = 3584  # MiB
 if RUNNING_ON_CLUSTER:
     SUITE = ["depot", "freecell", "gripper", "zenotravel"]
     ENVIRONMENT = BaselSlurmEnvironment(
@@ -66,20 +73,22 @@ ATTRIBUTES = [
     "error",
     "g_values_over_time",
     "run_dir",
+    "raw_memory",
     "runtime",
+    "virtual_memory",
 ]
 
 exp = Experiment(environment=ENVIRONMENT)
 exp.add_step("build", exp.build)
 exp.add_step("start", exp.start_runs)
 exp.add_fetcher(name="fetch")
-exp.add_parser(os.path.join(DIR, "singularity-parser.py"))
+exp.add_parser(DIR / "singularity-parser.py")
 
 
 def get_image(name):
     planner = name.replace("-", "_")
-    image = os.path.join(IMAGES_DIR, name + ".img")
-    assert os.path.exists(image), image
+    image = IMAGES_DIR / (name + ".img")
+    assert image.is_file(), image
     return planner, image
 
 
@@ -88,32 +97,47 @@ IMAGES = [get_image("fd1906-lama-first")]
 for planner, image in IMAGES:
     exp.add_resource(planner, image, symlink=True)
 
-singularity_script = os.path.join(DIR, "run-singularity.sh")
-exp.add_resource("run_singularity", singularity_script)
+exp.add_resource("run_singularity", DIR / "run-singularity.sh")
+exp.add_resource("filter_stderr", DIR / "filter-stderr.py")
 
 for planner, _ in IMAGES:
     for task in suites.build_suite(BENCHMARKS_DIR, SUITE):
         run = exp.add_run()
         run.add_resource("domain", task.domain_file, "domain.pddl")
         run.add_resource("problem", task.problem_file, "problem.pddl")
+        # Use runsolver to limit time and memory. It must be on the system
+        # PATH. Important: we cannot use time_limit and memory_limit of
+        # Lab's add_command() because setting the same memory limit with
+        # runsolver again using setrlimit fails.
         run.add_command(
             "run-planner",
             [
+                "runsolver",
+                "-C",
+                TIME_LIMIT,
+                "-V",
+                MEMORY_LIMIT,
+                "-w",
+                "watch.log",
+                "-v",
+                "values.log",
                 "{run_singularity}",
                 f"{{{planner}}}",
                 "{domain}",
                 "{problem}",
                 "sas_plan",
             ],
-            time_limit=TIME_LIMIT,
-            memory_limit=3584,
         )
+        # Remove temporary files from old Fast Downward versions.
+        run.add_command("rm-tmp-files", ["rm", "-f", "output.sas", "output"])
+        run.add_command("filter-stderr", [sys.executable, "{filter_stderr}"])
+
         run.set_property("domain", task.domain)
         run.set_property("problem", task.problem)
         run.set_property("algorithm", planner)
         run.set_property("id", [planner, task.domain, task.problem])
 
-report = os.path.join(exp.eval_dir, f"{exp.name}.html")
+report = Path(exp.eval_dir) / f"{exp.name}.html"
 exp.add_report(BaseReport(attributes=ATTRIBUTES), outfile=report)
 
 exp.run_steps()